chore[add query-for-package-url]: Updates UpdatePurlForPackageNamespaces function to use PkgToPurl function to construct purl
