+++ internal/testing/e2e/expectIsDependencyQ2.json
@@ -13,7 +13,7 @@
                 "name": "consul",
                 "versions": [
                   {
-                    "purl": "pkg:guac/cdx/docker.io/library/consul:sha256:22ab19cf1326abbfaafec6a14eb68f96e899e88ffe9ce26fa36affcf8ffb582c?tag=latest",
+                    "purl": "pkg:guac/cdx/docker.io/library/consul@sha256%3A22ab19cf1326abbfaafec6a14eb68f96e899e88ffe9ce26fa36affcf8ffb582c?tag=latest",
                     "version": "sha256:22ab19cf1326abbfaafec6a14eb68f96e899e88ffe9ce26fa36affcf8ffb582c",
                     "qualifiers": [
                       {
@@ -39,7 +39,7 @@
                 "name": "logrus",
                 "versions": [
                   {
-                    "purl": "pkg:golang/github.com/sirupsen/logrus:v1.4.2",
+                    "purl": "pkg:golang/github.com/sirupsen/logrus@v1.4.2",
                     "version": "v1.4.2",
                     "qualifiers": [],
                     "subpath": ""
