+++ internal/testing/e2e/expectPathPy.json
@@ -25,7 +25,7 @@
             "name": "python",
             "versions": [
               {
-                "purl": "pkg:guac/cdx/docker.io/library/python:sha256:e9ebb760d74df0c28447c2f67050800c3392a3987ed77ce493545c7d3ad24c9d?tag=latest",
+                "purl": "pkg:guac/cdx/docker.io/library/python@sha256%3Ae9ebb760d74df0c28447c2f67050800c3392a3987ed77ce493545c7d3ad24c9d?tag=latest",
                 "version": "sha256:e9ebb760d74df0c28447c2f67050800c3392a3987ed77ce493545c7d3ad24c9d",
                 "subpath": ""
               }
@@ -49,7 +49,7 @@
               "name": "python",
               "versions": [
                 {
-                  "purl": "pkg:guac/cdx/docker.io/library/python:sha256:e9ebb760d74df0c28447c2f67050800c3392a3987ed77ce493545c7d3ad24c9d?tag=latest",
+                  "purl": "pkg:guac/cdx/docker.io/library/python@sha256%3Ae9ebb760d74df0c28447c2f67050800c3392a3987ed77ce493545c7d3ad24c9d?tag=latest",
                   "version": "sha256:e9ebb760d74df0c28447c2f67050800c3392a3987ed77ce493545c7d3ad24c9d",
                   "subpath": ""
                 }
@@ -69,7 +69,7 @@
               "name": "libsqlite3-dev",
               "versions": [
                 {
-                  "purl": "pkg:deb/debian/libsqlite3-dev:3.34.1-3?arch=amd64&distro=debian-11&upstream=sqlite3",
+                  "purl": "pkg:deb/debian/libsqlite3-dev@3.34.1-3?arch=amd64&distro=debian-11&upstream=sqlite3",
                   "version": "3.34.1-3",
                   "subpath": ""
                 }
@@ -92,7 +92,7 @@
             "name": "libsqlite3-dev",
             "versions": [
               {
-                "purl": "pkg:deb/debian/libsqlite3-dev:3.34.1-3?arch=amd64&distro=debian-11&upstream=sqlite3",
+                "purl": "pkg:deb/debian/libsqlite3-dev@3.34.1-3?arch=amd64&distro=debian-11&upstream=sqlite3",
                 "version": "3.34.1-3",
                 "subpath": ""
               }
