+++ internal/testing/e2e/expectPkgQ3.json
@@ -10,7 +10,7 @@
               "name": "libp11-kit0",
               "versions": [
                 {
-                  "purl": "pkg:deb/debian/libp11-kit0:0.23.22-1?arch=amd64&distro=debian-11&upstream=p11-kit",
+                  "purl": "pkg:deb/debian/libp11-kit0@0.23.22-1?arch=amd64&distro=debian-11&upstream=p11-kit",
                   "version": "0.23.22-1",
                   "qualifiers": [
                     {
@@ -29,7 +29,7 @@
                   "subpath": ""
                 },
                 {
-                  "purl": "pkg:deb/debian/libp11-kit0:0.23.22-1?arch=arm64&distro=debian-11&upstream=p11-kit",
+                  "purl": "pkg:deb/debian/libp11-kit0@0.23.22-1?arch=arm64&distro=debian-11&upstream=p11-kit",
                   "version": "0.23.22-1",
                   "qualifiers": [
                     {
@@ -58,7 +58,7 @@
               "name": "libp11-kit0",
               "versions": [
                 {
-                  "purl": "pkg:deb/ubuntu/libp11-kit0:0.23.20-1ubuntu0.1?arch=amd64&distro=ubuntu-20.04&upstream=p11-kit",
+                  "purl": "pkg:deb/ubuntu/libp11-kit0@0.23.20-1ubuntu0.1?arch=amd64&distro=ubuntu-20.04&upstream=p11-kit",
                   "version": "0.23.20-1ubuntu0.1",
                   "qualifiers": [
                     {
@@ -77,7 +77,7 @@
                   "subpath": ""
                 },
                 {
-                  "purl": "pkg:deb/ubuntu/libp11-kit0:0.24.0-6build1?arch=amd64&distro=ubuntu-22.04&upstream=p11-kit",
+                  "purl": "pkg:deb/ubuntu/libp11-kit0@0.24.0-6build1?arch=amd64&distro=ubuntu-22.04&upstream=p11-kit",
                   "version": "0.24.0-6build1",
                   "qualifiers": [
                     {
