+++ pkg/assembler/helpers/purl_test.go
@@ -22,7 +22,6 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 	model "github.com/guacsec/guac/pkg/assembler/clients/generated"
-	assemblerModel "github.com/guacsec/guac/pkg/assembler/graphql/model"
 )
 
 var cmpOpts = []cmp.Option{
@@ -773,190 +772,4 @@ func TestGuacFilePurl(t *testing.T) {
 
 func strP(s string) *string {
 	return &s
-}
-
-var Package = &assemblerModel.Package{
-	Type: "conan",
-	Namespaces: []*assemblerModel.PackageNamespace{
-		{
-			Namespace: "openssl.org",
-			Names: []*assemblerModel.PackageName{
-				{
-					Name: "openssl",
-					Versions: []*assemblerModel.PackageVersion{
-						{
-							Version: "3.0.3",
-							Qualifiers: []*assemblerModel.PackageQualifier{
-								{
-									Key: "distro",
-									Value: "debian-11",
-								},
-							},
-						},
-					},
-				},
-				{
-					Name: "libp11-kit0",
-					Versions: []*assemblerModel.PackageVersion{
-						{
-							Version: "3.0.3",
-							Qualifiers: []*assemblerModel.PackageQualifier{
-								{
-									Key: "arch",
-									Value: "amd64",
-								  },
-								  {
-									Key: "distro",
-									Value: "debian-11",
-								  },
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			Namespace: "ubuntu",
-			Names: []*assemblerModel.PackageName{
-				{
-					Name: "openssl",
-					Versions: []*assemblerModel.PackageVersion{
-						{
-							Version: "3.0.3",
-							Qualifiers: []*assemblerModel.PackageQualifier{
-								{
-									Key: "arch",
-									Value: "amd64",
-								},
-							},
-							Subpath: "api",
-						},
-					},
-				},
-			},
-		},
-	},
-}
-
-var PackageNamespaces = []*assemblerModel.PackageNamespace{
-	{
-		Namespace: "openssl.org",
-		Names: []*assemblerModel.PackageName{
-			{
-				Name: "openssl",
-				Versions: []*assemblerModel.PackageVersion{
-					{
-						Purl:    "pkg:conan/openssl.org/openssl:3.0.3?distro=debian-11",
-						Version: "3.0.3",
-						Qualifiers: []*assemblerModel.PackageQualifier{
-							{
-								Key: "distro",
-								Value: "debian-11",
-							  },
-						},
-					},
-				},
-			},
-			{
-				Name: "libp11-kit0",
-				Versions: []*assemblerModel.PackageVersion{
-					{
-						Purl:    "pkg:conan/openssl.org/libp11-kit0:3.0.3?arch=amd64&distro=debian-11",
-						Version: "3.0.3",
-						Qualifiers: []*assemblerModel.PackageQualifier{
-							{
-								Key: "arch",
-								Value: "amd64",
-							  },
-							  {
-								Key: "distro",
-								Value: "debian-11",
-							  },
-						},
-					},
-				},
-			},
-		},
-	},
-	{
-		Namespace: "ubuntu",
-		Names: []*assemblerModel.PackageName{
-			{
-				Name: "openssl",
-				Versions: []*assemblerModel.PackageVersion{
-					{
-						Purl:    "pkg:conan/ubuntu/openssl:3.0.3#api?arch=amd64",
-						Version: "3.0.3",
-						Qualifiers: []*assemblerModel.PackageQualifier{
-							{
-								Key: "arch",
-								Value: "amd64",
-							},
-						},
-						Subpath: "api",
-					},
-				},
-			},
-		},
-	},
-}
-
-func TestUpdatePurlForNamespaces(t *testing.T) {
-	tests := []struct {
-		Name        string
-		Package     *assemblerModel.Package
-		ExpResult   []*assemblerModel.PackageNamespace
-		ExpQueryErr bool
-	}{
-		{
-			Name:        "Happy Path",
-			Package:     Package,
-			ExpResult:   PackageNamespaces,
-			ExpQueryErr: false,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.Name, func(t *testing.T) {
-			result, err := UpdatePurlForPackageNamespaces(test.Package)
-			if (err != nil) != test.ExpQueryErr {
-				t.Fatalf("did not get query error, want: %v, got: %v", test.ExpQueryErr, err)
-			}
-			if len(result) != len(test.ExpResult) {
-				t.Errorf("Length mismatch: expected %d, got %d", len(test.ExpResult), len(result))
-				return
-			}
-			fmt.Println(result)
-			for i := range result {
-				if !comparePackageNames(result[i].Names, test.ExpResult[i].Names) {
-					t.Errorf("Mismatch at index %d:", i)
-					fmt.Printf("Expected: %#v\n", test.ExpResult[i].Names)
-					fmt.Printf("Actual:   %#v\n", result[i].Names)
-				}
-			}
-		})
-	}
-
-}
-
-func comparePackageNames(packageNameOne, packageNameTwo []*assemblerModel.PackageName) bool {
-	if len(packageNameOne) != len(packageNameTwo) {
-		return false
-	}
-	for i := range packageNameOne {
-		if packageNameOne[i].Name != packageNameTwo[i].Name {
-			return false
-		}
-		if len(packageNameOne[i].Versions) != len(packageNameTwo[i].Versions) {
-			return false
-		}
-		for j := range packageNameOne[i].Versions {
-			if packageNameOne[i].Versions[j].Purl != packageNameTwo[i].Versions[j].Purl {
-				fmt.Printf("==%s", packageNameOne[i].Versions[j].Purl)
-				fmt.Printf("==%s", packageNameTwo[i].Versions[j].Purl)
-				return false
-			}
-		}
-	}
-	return true
 }
